fix: seperate controllers in to an individual position
