+++ server/routers/cinemas.mjs
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { getCinemasAll, getCinemasById } from "../controllers/controller.mjs";
+import { getCinemasAll, getCinemasById } from "../controllers/cinemas.controller.mjs";
 import { protect } from "../middlewares/protect.mjs";
 const cinemasRouter = Router();
 cinemasRouter.get("/", getCinemasAll);
