+++ server/routers/city.mjs
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { getCityAll, getCinemasByCity } from "../controllers/controller.mjs";
+import { getCityAll, getCinemasByCity } from "../controllers/city.controllers.mjs";
 import connectionPool from "../utils/db.mjs";
 import { protect } from "../middlewares/protect.mjs";
 // const express = require("express");
