+++ server/routers/days.mjs
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { getDaysAll } from "../controllers/controller.mjs";
+import { getDaysAll } from "../controllers/days.controller.mjs";
 const daysRouter = Router();
 daysRouter.get("/", getDaysAll);
