+++ server/routers/searchBar.mjs
@@ -1,5 +1,6 @@
 import { Router } from "express";
-import { getMoviesBySearchBar } from "../controllers/controller.mjs";
+// import { getMoviesBySearchBar } from "../controllers/controller.mjs";
+import { getMoviesBySearchBar } from "../controllers/searchbar.controller.mjs";
 const searchBarRouter = Router();
 
 searchBarRouter.get("/", getMoviesBySearchBar);
