+++ server/routers/ticket.mjs
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { getInfoForBookTicket } from "../controllers/controller.mjs";
+import { getInfoForBookTicket } from "../controllers/ticket.controller.mjs";
 const ticketRouter = Router();
 
 ticketRouter.get("/", getInfoForBookTicket);
